Create Design for Assignment 1
Design of Assignment 1 CPSC 351 
Molita Nuon 
Yocelin Esqueda
For this project, the design was created to use message passing and shared memory processes. 
Firstly, the receiver program created a shared memory and a message queue. The sender program then had access to the shared memory and stored bytes in the shared memory. Using the message queue, messages were sent and stored by the sender into the message queue, and they were received from that message queue by the receiver program. Thus, the receiver was able to obtain the bytes stored in the shared memory by the sender program. 
	
	
Message Passing:     		Shared memory: